refactor(gateway): rename types_legacy.rs → ts_signatures.rs, fix module doc [Q-L2/D9]
`types_legacy.rs` was misnamed: it is the LIVE TypeScript signature/.d.ts
generator called from types.rs, not backward-compat shims. Rename to
`ts_signatures.rs` and add a proper `//!` module doc explaining its role.

The old `types_legacy` module is kept as a thin re-export shim so any callers
using the old path continue to compile. `types.rs` is updated to call
`super::ts_signatures::generate_tool_types` directly. Tests are mirrored in
`tests_ts_signatures.rs` (with updated paths); `tests_types_legacy.rs` is kept
pointing through the compat shim to validate the alias works.

CLAUDE.md file-map row updated to reflect the true role of both files.
